separate visualizer source files
