+++ visualizer/main.cpp
@@ -1,359 +1,10 @@
-#include <SDL2/SDL.h>
-#include <SDL2/SDL_render.h>
-#include <SDL_ttf.h>
-#include <fmt/format.h>
-#include <imgui.h>
-#include <imgui_impl_sdl2.h>
-#include <imgui_impl_sdlrenderer2.h>
 #include <spdlog/spdlog.h>
 
-#include <algorithm>
 #include <argparse/argparse.hpp>
-#include <chrono>
-#include <qdpf.hpp>
-#include <string>
-#include <string_view>
-#include <unordered_map>
-#include <vector>
 
-enum Terrain {
-  Land = 0b001,
-  Water = 0b010,
-  Building = 0b100,
-};
+#include "visualizer.hpp"
 
-// Max value of w and h.
-const int N = 800;
-
-const int COST_UNIT = 100;
-
-using Cell = std::pair<int, int>;  // {x,y}
-
-struct CommandlineOptions {
-  int w = 100, h = 70;
-  // number of pixels per grid side.
-  int gridSize = 18;
-  // directory to fonts.
-  std::string fontsPath;
-  // step to pick gates, -1 for using stepf.
-  int step = -1;
-};
-
-// global options.
 CommandlineOptions options;
-
-struct Agent {
-  int size = COST_UNIT;
-  // the terrain capability for current agent.
-  int capability = Terrain::Land;
-
-  void Reset();
-};
-
-struct Map {
-  // qmx is never changed once set.
-  qdpf::QuadtreeMapX* qmx = nullptr;
-  // width and height (in cell/grids)
-  const int w = 32, h = 32;
-  // grid size in pixels
-  const int gridSize = 18;
-  // the value of GRIDS[x][y] is a terrain type integer.
-  int grids[N][N];
-  // CHANGES[x][y] stores the terrain value that cell (x,y) is going to change to.
-  // 0 for no changes.
-  int changes[N][N];
-  // step to pick gate cells.
-  int step = -1;
-
-  Map(int w, int h, int gridSize, int step = -1);
-  ~Map();
-  void Reset();
-  void BuildMapX();  // call only once.
-  // Change Terrain
-  void WantChangeTerrain(const Cell& cell, Terrain to);
-  void ApplyChangeTerrain(const std::vector<Cell>& cells);
-  void ClearAllTerrains();
-};
-
-struct AStarContext {
-  // never change once set.
-  qdpf::AStarPathFinder* pf = nullptr;
-  bool isPfReset = false;
-  // start cell
-  int x1 = 0, y1 = 0;
-  // target cell
-  int x2 = 0, y2 = 0;
-  // ~~~~~~ results ~~~~~~
-  std::vector<const qdpf::QdNode*> nodePath;
-  std::vector<Cell> gatePath;
-  std::vector<Cell> finalPath;
-
-  ~AStarContext();
-  void InitPf(qdpf::QuadtreeMapX* qmx);
-  void ClearResults();
-  void Reset();
-  int ResetPf(int agentSize, int capabilities);
-};
-
-template <typename V>
-struct FlowFieldItem {
-  V current;
-  V next;
-  int cost;
-};
-
-struct FlowFieldContext {
-  qdpf::FlowFieldPathFinder* pf = nullptr;
-  // target cell.
-  int x2 = 0, y2 = 0;
-  // query rectangle.
-  qdpf::Rectangle qrange;
-  // ~~~~~~ results ~~~~~~
-  std::vector<FlowFieldItem<const qdpf::QdNode*>> nodeFlowField;
-  std::vector<FlowFieldItem<Cell>> gateFlowField;
-  std::vector<FlowFieldItem<Cell>> finalFlowField;
-  std::unordered_map<Cell, Cell, qdpf::internal::PairHasher<int, int>> finalFlowNextMap;
-
-  bool isPfReset = false;
-
-  // ~~~~~ optional test path ~~~~~~
-  int x1 = -1, y1 = -1;  // start
-  std::vector<Cell> testPath;
-
-  ~FlowFieldContext();
-  void InitPf(qdpf::QuadtreeMapX* qmx);
-  int ResetPf(int agentSize, int capabilities);
-  void ClearResults();
-  void Reset();
-};
-
-// SDL coordinates, in pixels
-struct Camera {
-  // size of the camera window in pixels.
-  const int w, h;
-  // size of the whole  map in pixel.
-  const int mpw, mph;
-
-  // left corner (x,y), in pixel-level, on the whole map.
-  int x = 0, y = 0;
-  // changes.
-  int dx = 0, dy = 0;
-
-  Camera(int w, int h, int mpw, int mph);
-
-  // Movements
-  void MoveUp();
-  void MoveDown();
-  void MoveLeft();
-  void MoveRight();
-
-  // Update should be called after lots of MoveXX calls
-  void Update();
-};
-
-struct ArrowFont {
-  TTF_Font* font = nullptr;
-  SDL_Texture* texture;
-  // width and offset of each arrwo
-  int w[8];
-  int offset[8];
-  // font height
-  int h;
-};
-
-// char in arrows font.
-// 0 A →
-// 1 B ←
-// 2 C ↑
-// 3 D ↓
-// 4 E ↖
-// 5 F ↙
-// 6 G ↗
-// 7 H ↘
-const char ARROWS_CHAR[9] = "ABCDEFGH";
-
-// (dx+1)*3+(dy+1) => index in ARROWS_CHAR
-const int ARROWS_DIRECTIONS[9] = {
-    4,   // 0  ↖,
-    2,   // 1   ↑
-    6,   // 2   ↗
-    1,   // 3   ←
-    -1,  // NA
-    0,   // 5   →
-    5,   // 6   ↙
-    3,   // 7   ↓
-    7,   // 8   ↘,
-};
-
-// Interaction states.
-enum class State {
-  // Idle => AStarWaitStart | FlowFieldWaitQrangeLeftTop
-  Idle = 0,
-
-  DrawingBuildings = 1,
-  DrawingWaters = 2,
-
-  // ~~~~ AStar ~~~~
-  AStarWaitStart = 21,          // => AStarWaitTarget
-  AStarWaitTarget = 22,         // => AStarWaitCompution
-  AStarWaitCompution = 23,      // => AStarNodePathComputed  | AStarGatePathComputed
-  AStarNodePathComputed = 25,   // => AStarGatePathComputed
-  AStarGatePathComputed = 26,   // => AStarFinalPathComputed
-  AStarFinalPathComputed = 27,  // => Idle | AStarWaitStart
-
-  // ~~~~ FlowField ~~~~
-  FlowFieldWaitQrangeLeftTop = 31,      // => FlowFieldWaitQrangeRightBottom
-  FlowFieldWaitQrangeRightBottom = 32,  // => FlowFieldWaitTarget
-  FlowFieldWaitTarget = 33,             // => FlowFieldWaitCompution
-  FlowFieldWaitCompution = 34,       // => FlowFieldNodeLevelComputed | FlowFieldGateLevelComputed
-  FlowFieldNodeLevelComputed = 37,   // => FlowFieldGateLevelComputed
-  FlowFieldGateLevelComputed = 38,   // => FlowFieldFinalLevelComputed
-  FlowFieldFinalLevelComputed = 39,  // => Idle | FlowFieldWaitQrangeRightBottom
-};
-
-std::string StateToString(State state);
-
-enum class PathFinderFlag {
-  AStar = 0,
-  FlowField = 1,
-};
-
-const SDL_Color LightGray{180, 180, 180, 255};
-const SDL_Color Red{255, 0, 0, 255};
-const SDL_Color Blue{0, 0, 255, 255};
-const SDL_Color Black{0, 0, 0, 255};
-const SDL_Color Green{0, 180, 0, 255};
-const SDL_Color LightOrange{255, 200, 128, 255};
-const SDL_Color Orange{255, 165, 0, 255};
-const SDL_Color LightPurple{230, 190, 230, 255};
-
-const ImVec4 ImRed(0.7f, 0.2f, 0.2f, 1.0f);
-const ImVec4 ImWhite(1.0f, 1.0f, 1.0f, 1.0f);
-const ImVec4 ImGreen(0.0f, 1.0f, 0.0f, 1.0f);
-const ImVec4 ImBlue(0.2f, 0.2f, 0.7f, 1.0f);
-
-class Visualizer {
- public:
-  // w, h: the size of the grid map.
-  // gridSize: the size of one grid.
-  Visualizer(int w, int h, int gridSize, int step = -1);
-  int Init();
-  void Start();
-  void Destroy();
-  ~Visualizer();
-
- private:
-  State state = State::Idle;
-
-  SDL_Window* window = nullptr;
-  SDL_Renderer* renderer = nullptr;
-  Camera* camera = nullptr;
-
-  Agent agent;
-
-  Map map;
-
-  AStarContext astar;
-  FlowFieldContext flowfield;
-
-  PathFinderFlag pathfinderFlag = PathFinderFlag::AStar;
-
-  bool isMouseDown = false;
-
-  std::string messageHint = "";
-  ImVec4 messageHintColor = ImWhite;
-
-  bool stop = false;
-
-  // ~~~~~ changing terrains ~~~~~~
-  Terrain changeTo = Terrain::Building;
-  std::vector<Cell> changingTerrainCells;
-
-  // ~~~~~~ ui states ~~~~~~~
-  bool showClearAllTerrainsConfirm = false;
-
-  // ~~~~~~ imgui ~~~~~~~
-  ImFont* largeFont;
-
-  // ~~~~~~ misc ~~~~~~~
-  ArrowFont arrows;
-
-  void reset();
-  int initSDL();
-  int initImgui();
-  int initArrowsFont();
-  void destroyArrowsFont();
-  void destroyImgui();
-  void destroySDL();
-
-  // ~~~~~~ render the world ~~~~~~~
-  void renderWorld();
-  void renderGrids();
-  void renderQuadtreeNodes();
-  void renderGates();
-  void renderHighlightedNodes();
-  void renderHighlightedNodesAstar();
-  void renderHighlightedNodesFlowField();
-  void renderPathfindingDispatch();
-  void renderPathfindingAStar();
-  void renderPathfindingFlowField();
-  void renderPathFindingFlowFieldGateField();
-  void renderPathFindingFlowFieldFinalField();
-  void renderDrawRect(const SDL_Rect& rect, const SDL_Color& color);
-  void renderFillRect(const SDL_Rect& rect, const SDL_Color& color);
-  void renderDrawLine(int x1, int y1, int x2, int y2, const SDL_Color& color);
-  void renderCopy(SDL_Texture* texture, const SDL_Rect& src, const SDL_Rect& dst);
-  void renderFillCell(int x, int y, const SDL_Color& color);
-  void renderFillAgent(int x, int y);
-  void setMessageHint(std::string_view message, const ImVec4& color);
-  // ~~~~~~ render the panel ~~~~~~~
-  void renderImguiPanel();
-  void renderImguiPanelSectionCommon();
-  void renderImguiPanelSectionPathFinding();
-  void renderImguiPanelSectionPathFindingAStar();
-  void renderImguiPanelSectionPathFindingFlowField();
-  void renderImguiPanelSectionAgent();
-  // ~~~~ camera ~~~~~
-  bool cropRectByCamera(const SDL_Rect& rect, SDL_Rect& overlap,
-                        bool marginToCameraCoordinates = true);
-  // ~~~~ astar ~~~~~~~
-  void computeAstarNodePath();
-  void computeAstarGatePath();
-  void computeAstarFinalPath();
-  void handleAstarInputBegin();
-  void handleFlowFieldInputQueryRangeBegin();
-  // ~~~~~ flowfield ~~~~~~
-  void computeNodeFlowField();
-  void computeGateFlowField();
-  void computeFinalFlowField();
-  void handlePlayFlowFieldTestPath();
-  // ~~~~ interaction ~~~~~~~
-  void handleInputs();
-  void handleInputsShortcuts(SDL_Event& e);
-  void handleInputsForCrameMovementsByKeyBoard(SDL_Event& e);
-  void handleInputsForCrameMovementsByMouse(SDL_Event& e);
-  void handleInputsDispatchByState(SDL_Event& e);
-  void handleInputsChangeTerrains(SDL_Event& e);
-  void handleInputsAStarSetStart(SDL_Event& e);
-  void handleInputsAStarSetTarget(SDL_Event& e);
-  void handleInputsFlowFieldSetQrangeLeftTop(SDL_Event& e);
-  void handleInputsFlowFieldSetQrangeRightBottom(SDL_Event& e);
-  void handleInputsFlowFieldSetTarget(SDL_Event& e);
-  void handleInputsFlowFieldSetTestStart(SDL_Event& e);
-  // ~~~~~~~~ handlers ~~~~~~~~~~~
-  void handleLogics();
-  void handleStartDrawBuildings();
-  void handleStartDrawWater();
-  void pushTerrainChanges(const Cell&);
-  void applyTerrainChanges();
-  void handleSwitchPathFinderHandler(PathFinderFlag to);
-  void handleChangeAgentSize(int to);
-  void handleChangeAgentCompability(int to);
-  void handleClearAllTerrains();
-  // ~~~~~~ util ~~~~~~~
-  std::pair<int, int> getCellAtPixelPosition(int x, int y) const;
-  const qdpf::internal::QuadtreeMap* getCurrentQuadtreeMapByAgent() const;
-};
 
 // Parse options from command line.
 // Returns 0 on success.
@@ -366,1498 +17,6 @@ int main(int argc, char* argv[]) {
   visualizer.Start();
   visualizer.Destroy();
   return 0;
-}
-
-// ~~~~~~~~~~~~~~~~~~~~ Implements ~~~~~~~~~~~~~~~~~~~~~~
-
-void Agent::Reset() {
-  size = COST_UNIT;
-  capability = Terrain::Land;
-}
-
-Map::Map(int w, int h, int gridSize, int step) : w(w), h(h), gridSize(gridSize), step(step) {
-  Reset();
-}
-
-Map::~Map() {
-  delete qmx;
-  qmx = nullptr;
-}
-
-void Map::Reset() {
-  // Inits the map, by default:
-  // 1. the center are water.
-  // 2. there's 2 walls.
-  // 3. other cells are land.
-  memset(grids, 0, sizeof grids);
-  memset(changes, 0, sizeof changes);
-
-  qdpf::Rectangle center{2 * h / 5, 2 * w / 5, h * 3 / 5, w * 3 / 5};
-  qdpf::Rectangle wall1{center.x2 + 4, center.y1, center.x2 + 4, center.y2 + 4};
-  qdpf::Rectangle wall2{center.x1, center.y2 + 4, center.x2 + 4, center.y2 + 4};
-  qdpf::Rectangle wall3{center.x1 - 4, center.y1 - 4, center.x1 - 4, center.y2};
-  qdpf::Rectangle wall4{center.x1 - 4, center.y1 - 4, center.x2, center.y1 - 4};
-
-  for (int x = 0; x < h; ++x) {
-    for (int y = 0; y < w; ++y) {
-      if (x >= center.x1 && x <= center.x2 && y >= center.y1 && y <= center.y2)
-        grids[x][y] = Terrain::Water;
-      else if (x >= wall1.x1 && x <= wall1.x2 && y >= wall1.y1 && y <= wall1.y2)
-        grids[x][y] = Terrain::Building;
-      else if (x >= wall2.x1 && x <= wall2.x2 && y >= wall2.y1 && y <= wall2.y2)
-        grids[x][y] = Terrain::Building;
-      else if (x >= wall3.x1 && x <= wall3.x2 && y >= wall3.y1 && y <= wall3.y2)
-        grids[x][y] = Terrain::Building;
-      else if (x >= wall4.x1 && x <= wall4.x2 && y >= wall4.y1 && y <= wall4.y2)
-        grids[x][y] = Terrain::Building;
-      else
-        grids[x][y] = Terrain::Land;
-    }
-  }
-}
-
-void Map::BuildMapX() {
-  auto stepf = (step == -1) ? [](int z) -> int { return z / 8 + 1; } : nullptr;
-  qdpf::QuadtreeMapXSettings settings{
-      {COST_UNIT, Terrain::Land},
-      {2 * COST_UNIT, Terrain::Land},
-      {3 * COST_UNIT, Terrain::Land},
-      {COST_UNIT, Terrain::Water},
-      {2 * COST_UNIT, Terrain::Water},
-      {3 * COST_UNIT, Terrain::Water},
-      {COST_UNIT, Terrain::Land | Terrain::Water},
-      {2 * COST_UNIT, Terrain::Land | Terrain::Water},
-      {3 * COST_UNIT, Terrain::Land | Terrain::Water},
-  };
-  qmx = new qdpf::QuadtreeMapX(
-      w, h, qdpf::EuclideanDistance<COST_UNIT>, [this](int x, int y) { return grids[x][y]; },
-      settings, step, stepf);
-  qmx->Build();
-}
-
-void Map::WantChangeTerrain(const Cell& cell, Terrain to) {
-  auto [x, y] = cell;
-  changes[x][y] = to;
-}
-
-void Map::ApplyChangeTerrain(const std::vector<Cell>& cells) {
-  for (auto [x, y] : cells) {
-    grids[x][y] = changes[x][y];
-    changes[x][y] = 0;
-    qmx->Update(x, y);
-  }
-  qmx->Compute();
-}
-
-void Map::ClearAllTerrains() {
-  for (int x = 0; x < h; ++x) {
-    for (int y = 0; y < w; ++y) {
-      grids[x][y] = Terrain::Land;
-      qmx->Update(x, y);
-    }
-  }
-  qmx->Compute();
-}
-
-void AStarContext::InitPf(qdpf::QuadtreeMapX* qmx) { pf = new qdpf::AStarPathFinder(*qmx); }
-
-AStarContext::~AStarContext() {
-  delete pf;
-  pf = nullptr;
-}
-
-void AStarContext::ClearResults() { nodePath.clear(), gatePath.clear(), finalPath.clear(); }
-
-void AStarContext::Reset() {
-  ClearResults();
-  x1 = y1 = x2 = y2 = 0;
-  isPfReset = false;
-}
-
-int AStarContext::ResetPf(int agentSize, int capabilities) {
-  if (isPfReset) return 0;
-  auto ret = pf->Reset(x1, y1, x2, y2, agentSize, capabilities);
-  isPfReset = true;
-  return ret;
-}
-
-FlowFieldContext::~FlowFieldContext() {
-  delete pf;
-  pf = nullptr;
-}
-
-void FlowFieldContext::InitPf(qdpf::QuadtreeMapX* qmx) {
-  pf = new qdpf::FlowFieldPathFinder(*qmx);
-}
-
-int FlowFieldContext::ResetPf(int agentSize, int capabilities) {
-  if (isPfReset) return 0;
-  auto ret = pf->Reset(x2, y2, qrange, agentSize, capabilities);
-  isPfReset = true;
-  return ret;
-}
-
-void FlowFieldContext::ClearResults() {
-  nodeFlowField.clear(), gateFlowField.clear(), finalFlowField.clear();
-  finalFlowNextMap.clear();
-  testPath.clear();
-}
-
-void FlowFieldContext::Reset() {
-  ClearResults();
-  x2 = y2 = 0;
-  x1 = y1 = -1;
-  qrange = {0, 0, 0, 0};
-  isPfReset = false;
-}
-
-Camera::Camera(int w, int h, int mpw, int mph) : w(w), h(h), mpw(mpw), mph(mph) {}
-
-void Camera::MoveUp() { dy -= 50; }
-void Camera::MoveDown() { dy += 50; }
-void Camera::MoveLeft() { dx -= 50; }
-void Camera::MoveRight() { dx += 50; }
-
-void Camera::Update() {
-  if (dx == 0 && dy == 0) return;
-  x += dx;
-  y += dy;
-  // bounds check.
-  x = std::max(0, x);
-  y = std::max(0, y);
-  x = std::min(x, mpw - w);
-  y = std::min(y, mph - h);
-  // clears the delta changes.
-  dx = dy = 0;
-}
-
-Visualizer::Visualizer(int w, int h, int gridSize, int step) : map(Map(w, h, gridSize, step)) {}
-
-Visualizer::~Visualizer() {
-  delete camera;
-  camera = nullptr;
-}
-
-int Visualizer::Init() {
-  if (0 != initSDL()) return -1;
-  if (0 != initImgui()) {
-    destroySDL();
-    return -1;
-  }
-  if (0 != initArrowsFont()) {
-    destroyImgui();
-    destroySDL();
-    return -1;
-  }
-  spdlog::info("Visualizer Init done.");
-
-  spdlog::info("Building map (may take some time...)");
-  // Build the map.
-  map.BuildMapX();
-  // Build the pfs;
-  astar.InitPf(map.qmx);
-  flowfield.InitPf(map.qmx);
-  spdlog::info("Building map done");
-  return 0;
-}
-
-int Visualizer::initSDL() {
-  // Init SDL.
-  if (SDL_Init(SDL_INIT_EVERYTHING) != 0) {
-    spdlog::error("Error: {}", SDL_GetError());
-    return -1;
-  }
-
-  // Init ttf font
-  if (TTF_Init() == -1) {
-    spdlog::error("SDL_ttf Error: {}", SDL_GetError());
-    SDL_Quit();
-    return -1;
-  }
-
-  // Get display bounds
-  SDL_Rect displayBounds;
-  if (SDL_GetDisplayBounds(0, &displayBounds) != 0) {
-    spdlog::error("Failed to get display bounds: {}", SDL_GetError());
-    TTF_Quit();
-    SDL_Quit();
-    return 1;
-  }
-
-  // Creates window and camera.
-  int mpw = map.w * map.gridSize, mph = map.h * map.gridSize;
-  int w = std::min(mpw, displayBounds.w);
-  int h = std::min(mph, displayBounds.h);
-  window =
-      SDL_CreateWindow("quadtree-pathfinding-visualizer", SDL_WINDOWPOS_CENTERED,
-                       SDL_WINDOWPOS_CENTERED, w, h, SDL_WINDOW_SHOWN | SDL_WINDOW_ALLOW_HIGHDPI);
-  if (window == nullptr) {
-    spdlog::error("Create window error: {}", SDL_GetError());
-    TTF_Quit();
-    SDL_Quit();
-    return -3;
-  }
-  camera = new Camera(w, h, mpw, mph);
-
-  // Creates renderer.
-  renderer = SDL_CreateRenderer(window, -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC);
-  if (renderer == nullptr) {
-    spdlog::error("Create renderer error: {}", SDL_GetError());
-    SDL_DestroyWindow(window);
-    TTF_Quit();
-    SDL_Quit();
-    return -1;
-  }
-  return 0;
-}
-
-int Visualizer::initImgui() {
-  IMGUI_CHECKVERSION();
-  ImGui::CreateContext();
-  auto& io = ImGui::GetIO();
-  io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;  // Enable Keyboard Controls
-  io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;   // Enable Gamepad Controls
-  io.Fonts->AddFontFromFileTTF((options.fontsPath + "/Roboto-Medium.ttf").c_str(), 18);
-  largeFont = io.Fonts->AddFontFromFileTTF((options.fontsPath + "/Roboto-Medium.ttf").c_str(), 24);
-  // Setup Dear ImGui style
-  ImGui::StyleColorsDark();
-  // ImGui::StyleColorsLight();
-  // Setup Platform/Renderer backends
-  ImGui_ImplSDL2_InitForSDLRenderer(window, renderer);
-  ImGui_ImplSDLRenderer2_Init(renderer);
-  return 0;
-}
-
-int Visualizer::initArrowsFont() {
-  // Open font arrows.
-  std::string arrowsFontPath = options.fontsPath + "/Arrows.ttf";
-  arrows.font = TTF_OpenFont(arrowsFontPath.c_str(), 18);
-  if (arrows.font == nullptr) {
-    spdlog::error("Cannot open Arrows.ttf: {} {}", arrowsFontPath, SDL_GetError());
-    return -1;
-  }
-
-  // Creates surface for arrows font.
-  SDL_Surface* ts = TTF_RenderUTF8_Solid(arrows.font, ARROWS_CHAR, {0, 0, 0, 255});
-  if (!ts) {
-    spdlog::error("SDL_Surface: {}", TTF_GetError());
-    TTF_CloseFont(arrows.font);
-    return -1;
-  }
-
-  // Create arrows font texture.
-  arrows.texture = SDL_CreateTextureFromSurface(renderer, ts);
-  if (arrows.texture == nullptr) {
-    spdlog::error("Create arrows font texture: {}", SDL_GetError());
-    SDL_FreeSurface(ts);
-    TTF_CloseFont(arrows.font);
-    return -1;
-  }
-
-  // Computes the font height, weight and offset for each char.
-  int offset = 0;
-  for (int i = 0; i < 8; i++) {
-    int minx, maxx, miny, maxy, advance;
-    TTF_GlyphMetrics(arrows.font, ARROWS_CHAR[i], &minx, &maxx, &miny, &maxy, &advance);
-    arrows.w[i] = advance;
-    arrows.offset[i] = offset;
-    offset += advance;
-  }
-  arrows.h = TTF_FontHeight(arrows.font);
-
-  SDL_FreeSurface(ts);
-  return 0;
-}
-
-void Visualizer::Destroy() {
-  destroyArrowsFont();
-  destroyImgui();
-  destroySDL();
-}
-
-void Visualizer::destroyArrowsFont() {
-  if (arrows.font != nullptr) {
-    TTF_CloseFont(arrows.font);
-    arrows.font = nullptr;
-  }
-  if (arrows.texture != nullptr) {
-    SDL_DestroyTexture(arrows.texture);
-    arrows.texture = nullptr;
-  }
-}
-
-void Visualizer::destroySDL() {
-  // deinit SDL.
-  if (renderer) SDL_DestroyRenderer(renderer);
-  if (window) SDL_DestroyWindow(window);
-  if (arrows.font != nullptr) TTF_CloseFont(arrows.font);
-  TTF_Quit();
-  SDL_Quit();
-}
-
-void Visualizer::destroyImgui() {
-  // deinit ImGui.
-  ImGui_ImplSDLRenderer2_Shutdown();
-  ImGui_ImplSDL2_Shutdown();
-  ImGui::DestroyContext();
-}
-
-void Visualizer::Start() {
-  auto& io = ImGui::GetIO();
-  while (!stop) {
-    // quit on -1
-    handleInputs();
-    if (stop) break;
-
-    // update camera.
-    if (camera != nullptr) camera->Update();
-
-    handleLogics();
-
-    // starting a new imgui frame
-    ImGui_ImplSDLRenderer2_NewFrame();
-    ImGui_ImplSDL2_NewFrame();
-    ImGui::NewFrame();
-
-    // render imgui panel to imgui's buffer.
-    renderImguiPanel();
-    ImGui::Render();
-
-    // clears SDL buffer (white background)
-    SDL_SetRenderDrawColor(renderer, 255, 255, 255, 255);
-    SDL_RenderClear(renderer);
-
-    // render the world
-    renderWorld();
-
-    // pass the imgui's render buffer to SDL's renderer.
-    SDL_RenderSetScale(renderer, io.DisplayFramebufferScale.x, io.DisplayFramebufferScale.y);
-    ImGui_ImplSDLRenderer2_RenderDrawData(ImGui::GetDrawData(), renderer);
-
-    // present sdl buffer.
-    SDL_RenderPresent(renderer);
-    SDL_Delay(50);  // ms
-  }
-}
-
-void Visualizer::handleInputs() {
-  SDL_Event e;
-  auto& io = ImGui::GetIO();
-  while (SDL_PollEvent(&e)) {
-    ImGui_ImplSDL2_ProcessEvent(&e);
-    if (e.type == SDL_QUIT) {
-      stop = true;
-      return;
-    }
-
-    // isMouseDown;
-    if (e.type == SDL_MOUSEBUTTONDOWN)
-      isMouseDown = true;
-    else if (e.type == SDL_MOUSEBUTTONUP)
-      isMouseDown = false;
-
-    // shortcuts (in front of imgui)
-    handleInputsShortcuts(e);
-    handleInputsForCrameMovementsByKeyBoard(e);
-
-    // ImGui didn't handle this event, turn it to SDL.
-    if (!io.WantCaptureMouse && !io.WantCaptureKeyboard) {
-      handleInputsForCrameMovementsByMouse(e);
-      handleInputsDispatchByState(e);
-    }
-  }
-}
-
-// returns -1 for stop the whole window.
-void Visualizer::handleInputsShortcuts(SDL_Event& e) {
-  // shortcuts
-  if (e.type == SDL_KEYDOWN) {
-    switch (e.key.keysym.sym) {
-        // Ctrl-C
-      case SDLK_c:
-        if (SDL_GetModState() & KMOD_CTRL) stop = true;
-        break;
-        // ESC
-      case SDLK_ESCAPE:
-        reset();
-        break;
-      case SDLK_b:
-        handleStartDrawBuildings();
-        break;
-      case SDLK_w:
-        handleStartDrawWater();
-        break;
-    }
-  }
-}
-
-void Visualizer::handleInputsForCrameMovementsByMouse(SDL_Event& e) {
-  switch (e.type) {
-    case SDL_MOUSEWHEEL:
-      if (e.wheel.x > 0) camera->MoveLeft();
-      if (e.wheel.x < 0) camera->MoveRight();
-      if (e.wheel.y > 0) camera->MoveUp();
-      if (e.wheel.y < 0) camera->MoveDown();
-      break;
-  }
-}
-
-void Visualizer::handleInputsForCrameMovementsByKeyBoard(SDL_Event& e) {
-  switch (e.type) {
-    case SDL_KEYDOWN:
-      if (e.key.keysym.sym == SDLK_UP || e.key.keysym.sym == SDLK_k) camera->MoveUp();
-      if (e.key.keysym.sym == SDLK_DOWN || e.key.keysym.sym == SDLK_j) camera->MoveDown();
-      if (e.key.keysym.sym == SDLK_LEFT || e.key.keysym.sym == SDLK_h) camera->MoveLeft();
-      if (e.key.keysym.sym == SDLK_RIGHT || e.key.keysym.sym == SDLK_l) camera->MoveRight();
-      break;
-  }
-}
-
-void Visualizer::handleInputsDispatchByState(SDL_Event& e) {
-  switch (state) {
-    case State::DrawingBuildings:
-      [[fallthrough]];
-    case State::DrawingWaters:
-      handleInputsChangeTerrains(e);
-      break;
-    case State::AStarWaitStart:
-      handleInputsAStarSetStart(e);
-      break;
-    case State::AStarWaitTarget:
-      handleInputsAStarSetTarget(e);
-      break;
-    case State::FlowFieldWaitQrangeLeftTop:
-      handleInputsFlowFieldSetQrangeLeftTop(e);
-      break;
-    case State::FlowFieldWaitQrangeRightBottom:
-      handleInputsFlowFieldSetQrangeRightBottom(e);
-      break;
-    case State::FlowFieldWaitTarget:
-      handleInputsFlowFieldSetTarget(e);
-      break;
-    case State::FlowFieldFinalLevelComputed:
-      // reset test path's star:
-      handleInputsFlowFieldSetTestStart(e);
-      break;
-    default:
-      break;  // avoid warning.
-  }
-}
-
-void Visualizer::handleInputsChangeTerrains(SDL_Event& e) {
-  switch (e.type) {
-    case SDL_MOUSEBUTTONUP:
-      applyTerrainChanges();
-      break;
-    case SDL_MOUSEMOTION:
-      if (isMouseDown) pushTerrainChanges(getCellAtPixelPosition(e.button.x, e.button.y));
-      break;
-    case SDL_MOUSEBUTTONDOWN:
-      pushTerrainChanges(getCellAtPixelPosition(e.button.x, e.button.y));
-      break;
-  }
-}
-
-void Visualizer::handleInputsAStarSetStart(SDL_Event& e) {
-  if (e.type == SDL_MOUSEBUTTONDOWN) {
-    auto cell = getCellAtPixelPosition(e.button.x, e.button.y);
-    astar.x1 = cell.first;
-    astar.y1 = cell.second;
-    state = State::AStarWaitTarget;
-    setMessageHint("A*: waiting to click a target cell", ImGreen);
-  }
-}
-
-void Visualizer::handleInputsAStarSetTarget(SDL_Event& e) {
-  if (e.type == SDL_MOUSEBUTTONDOWN) {
-    auto cell = getCellAtPixelPosition(e.button.x, e.button.y);
-    astar.x2 = cell.first;
-    astar.y2 = cell.second;
-    state = State::AStarWaitCompution;
-    setMessageHint("A*: waiting to click a compution button", ImGreen);
-  }
-}
-
-void Visualizer::handleInputsFlowFieldSetQrangeLeftTop(SDL_Event& e) {
-  if (e.type == SDL_MOUSEBUTTONDOWN) {
-    auto cell = getCellAtPixelPosition(e.button.x, e.button.y);
-    flowfield.qrange.x1 = cell.first;
-    flowfield.qrange.y1 = cell.second;
-    state = State::FlowFieldWaitQrangeRightBottom;
-    setMessageHint("FlowField: waiting to click a right-bottom cell", ImGreen);
-  }
-}
-void Visualizer::handleInputsFlowFieldSetQrangeRightBottom(SDL_Event& e) {
-  if (e.type == SDL_MOUSEBUTTONDOWN) {
-    auto cell = getCellAtPixelPosition(e.button.x, e.button.y);
-    flowfield.qrange.x2 = cell.first;
-    flowfield.qrange.y2 = cell.second;
-    state = State::FlowFieldWaitTarget;
-    setMessageHint("FlowField: waiting to click a target cell", ImGreen);
-  }
-}
-
-void Visualizer::handleInputsFlowFieldSetTarget(SDL_Event& e) {
-  if (e.type == SDL_MOUSEBUTTONDOWN) {
-    auto cell = getCellAtPixelPosition(e.button.x, e.button.y);
-    flowfield.x2 = cell.first;
-    flowfield.y2 = cell.second;
-    state = State::FlowFieldWaitCompution;
-    setMessageHint("FlowField: waiting to click a compution button", ImGreen);
-  }
-}
-
-void Visualizer::handleInputsFlowFieldSetTestStart(SDL_Event& e) {
-  if (e.type == SDL_MOUSEBUTTONDOWN) {
-    auto cell = getCellAtPixelPosition(e.button.x, e.button.y);
-    auto [x, y] = cell;
-    if (flowfield.finalFlowNextMap.find({x, y}) == flowfield.finalFlowNextMap.end()) {
-      setMessageHint("FlowField: test path start invalid", ImRed);
-      flowfield.x1 = -1;
-      flowfield.y1 = -1;
-      return;
-    }
-    flowfield.x1 = x;
-    flowfield.y1 = y;
-    flowfield.testPath.clear();
-    setMessageHint("FlowField: playing a test path ...", ImGreen);
-  }
-}
-
-void Visualizer::renderImguiPanel() {
-  // window's background
-  ImGui::PushStyleColor(ImGuiCol_WindowBg, ImVec4(0.0f, 0.0f, 0.0f, 0.7f));
-
-  ImGui::Begin("Quadtree Pathfinder Visualizer");
-
-  // Status
-  ImGui::PushStyleColor(ImGuiCol_Text, ImWhite);
-  ImGui::PushStyleColor(ImGuiCol_HeaderHovered, ImBlue);
-  ImGui::PushFont(largeFont);
-  ImGui::Selectable(fmt::format("Status: {}", StateToString(state)).c_str(), true);
-  ImGui::PopFont();
-  ImGui::PopStyleColor(2);
-
-  ImGui::Spacing();
-
-  // Hint
-  ImGui::PushStyleColor(ImGuiCol_Text, messageHintColor);
-  ImGui::TextWrapped("HINT: %s", messageHint.c_str());
-  ImGui::PopStyleColor();
-  ImGui::Spacing();
-
-  // Sections.
-  renderImguiPanelSectionCommon();
-  ImGui::Spacing();
-  renderImguiPanelSectionAgent();
-  ImGui::Spacing();
-  renderImguiPanelSectionPathFinding();
-
-  ImGui::End();
-
-  ImGui::PopStyleColor();
-}
-
-void Visualizer::renderImguiPanelSectionCommon() {
-  ImGui::SeparatorText("Common");
-
-  // Reset.
-  if (ImGui::Button("Reset < ESC >")) reset();
-  if (ImGui::IsItemHovered()) {
-    ImGui::BeginTooltip();
-    ImGui::Text("Clear the compution results and back to idle. (key: ESC)");
-    ImGui::EndTooltip();
-  }
-  ImGui::SameLine();
-
-  // Button: Buildings
-  ImGui::PushStyleColor(ImGuiCol_Button, ImRed);
-  if (ImGui::Button("Building < B > ")) {
-    handleStartDrawBuildings();
-  }
-  ImGui::PopStyleColor(3);  // rollback color
-  if (ImGui::IsItemHovered()) {
-    ImGui::BeginTooltip();
-    ImGui::Text("Add or Remove building");
-    ImGui::EndTooltip();
-  }
-
-  ImGui::SameLine();
-
-  // Button: Water
-  ImGui::PushStyleColor(ImGuiCol_Button, ImBlue);
-  if (ImGui::Button("Water < W >")) {
-    handleStartDrawWater();
-  }
-  ImGui::PopStyleColor(3);  // rollback color
-  if (ImGui::IsItemHovered()) {
-    ImGui::BeginTooltip();
-    ImGui::Text("Add or Remove water");
-    ImGui::EndTooltip();
-  }
-
-  ImGui::SameLine();
-
-  // Button clear terrains;
-  if (ImGui::Button("Clear All Building and Water")) {
-    showClearAllTerrainsConfirm = true;
-    ImGui::OpenPopup("ClearAllTerrainsConfirmationDialog");
-  }
-  if (ImGui::IsItemHovered()) {
-    ImGui::BeginTooltip();
-    ImGui::Text("Clear all building and water on the map");
-    ImGui::EndTooltip();
-  }
-  if (showClearAllTerrainsConfirm) {
-    if (ImGui::BeginPopupModal("ClearAllTerrainsConfirmationDialog", &showClearAllTerrainsConfirm,
-                               ImGuiWindowFlags_AlwaysAutoResize)) {
-      ImGui::Text("Are you sure to clear all terrains to land ?");
-      ImGui::Separator();
-      if (ImGui::Button("Yes")) {
-        handleClearAllTerrains();
-        showClearAllTerrainsConfirm = false;
-        ImGui::CloseCurrentPopup();
-      }
-      ImGui::SetItemDefaultFocus();
-      ImGui::SameLine();
-      if (ImGui::Button("No")) {
-        showClearAllTerrainsConfirm = false;
-        ImGui::CloseCurrentPopup();
-      }
-
-      ImGui::EndPopup();
-    }
-  }
-}
-
-void Visualizer::renderImguiPanelSectionAgent() {
-  ImGui::SeparatorText("Agent");
-
-  ImGui::Text("Size: ");
-  ImGui::SameLine();
-  if (ImGui::RadioButton("1x1", agent.size == 1 * COST_UNIT)) handleChangeAgentSize(COST_UNIT);
-  ImGui::SameLine();
-  if (ImGui::RadioButton("2x2", agent.size == 2 * COST_UNIT)) handleChangeAgentSize(2 * COST_UNIT);
-  ImGui::SameLine();
-  if (ImGui::RadioButton("3x3", agent.size == 3 * COST_UNIT)) handleChangeAgentSize(3 * COST_UNIT);
-
-  ImGui::Text("Capability: ");
-  ImGui::SameLine();
-  if (ImGui::RadioButton("Land", agent.capability == Terrain::Land))
-    handleChangeAgentCompability(Terrain::Land);
-  ImGui::SameLine();
-  if (ImGui::RadioButton("Water", agent.capability == Terrain::Water))
-    handleChangeAgentCompability(Terrain::Water);
-  ImGui::SameLine();
-  if (ImGui::RadioButton("Land | Water", agent.capability == (Terrain::Land | Terrain::Water)))
-    handleChangeAgentCompability(Terrain::Land | Terrain::Water);
-}
-
-void Visualizer::renderImguiPanelSectionPathFinding() {
-  ImGui::SeparatorText("Pathfinding");
-
-  if (ImGui::RadioButton("A*", pathfinderFlag == PathFinderFlag::AStar))
-    handleSwitchPathFinderHandler(PathFinderFlag::AStar);
-  ImGui::SameLine();
-  if (ImGui::RadioButton("FlowField", pathfinderFlag == PathFinderFlag::FlowField))
-    handleSwitchPathFinderHandler(PathFinderFlag::FlowField);
-
-  switch (pathfinderFlag) {
-    case PathFinderFlag::AStar:
-      return renderImguiPanelSectionPathFindingAStar();
-    case PathFinderFlag::FlowField:
-      return renderImguiPanelSectionPathFindingFlowField();
-  }
-}
-
-void Visualizer::renderImguiPanelSectionPathFindingAStar() {
-  if (ImGui::Button("Set Start and Target")) {
-    handleAstarInputBegin();
-  }
-  ImGui::SameLine();
-  if (ImGui::Button("Compute Node Path")) {
-    computeAstarNodePath();
-  }
-  ImGui::SameLine();
-  if (ImGui::Button("Compute Gate Path")) {
-    computeAstarGatePath();
-  }
-  ImGui::SameLine();
-  if (ImGui::Button("Compute Final Path")) {
-    computeAstarFinalPath();
-  }
-}
-
-void Visualizer::renderImguiPanelSectionPathFindingFlowField() {
-  if (ImGui::Button("Set Start Rectangle and Target")) {
-    handleFlowFieldInputQueryRangeBegin();
-  }
-
-  if (ImGui::IsItemHovered()) {
-    ImGui::BeginTooltip();
-    ImGui::Text(
-        "Click the left-top and right-bottom corners of the start rectangle, and then click a "
-        "target");
-    ImGui::EndTooltip();
-  }
-
-  ImGui::SameLine();
-
-  if (ImGui::Button("Compute Node FlowField")) {
-    computeNodeFlowField();
-  }
-
-  ImGui::SameLine();
-
-  if (ImGui::Button("Compute Gate FlowField")) {
-    computeGateFlowField();
-  }
-
-  ImGui::SameLine();
-
-  if (ImGui::Button("Compute Final FlowField")) {
-    computeFinalFlowField();
-  }
-}
-
-void Visualizer::renderWorld() {
-  renderHighlightedNodes();
-  renderGates();
-  renderGrids();
-  renderQuadtreeNodes();
-  renderPathfindingDispatch();
-}
-
-void Visualizer::renderGrids() {
-  // (i,j) is the cell's position.
-  // (x,y) is the grid in pixel (in SDL coordinates).
-  for (int i = 0, y = 0; i < map.h; ++i, y += map.gridSize) {
-    for (int j = 0, x = 0; j < map.w; ++j, x += map.gridSize) {
-      SDL_Rect grid = {x, y, map.gridSize, map.gridSize};
-      SDL_Rect gridInner = {grid.x + 1, grid.y + 1, grid.w - 2, grid.h - 2};
-      renderDrawRect(grid, LightGray);
-      // Its terrain value or changing to value
-      auto v = map.grids[i][j];
-      if (map.changes[i][j] != 0) v = map.changes[i][j];
-      switch (v) {
-        case Terrain::Building:
-          renderFillRect(gridInner, Red);
-          break;
-        case Terrain::Water:
-          renderFillRect(gridInner, Blue);
-          break;
-      }
-    }
-  }
-}
-
-// render quadtree nodes of current map.
-void Visualizer::renderQuadtreeNodes() {
-  auto mp = getCurrentQuadtreeMapByAgent();
-  if (mp != nullptr) {
-    qdpf::internal::QdNodeVisitor c1 = [this](const qdpf::internal::QdNode* node) {
-      int x = node->y1 * map.gridSize;
-      int y = node->x1 * map.gridSize;
-      int w = (node->y2 - node->y1 + 1) * map.gridSize;
-      int h = (node->x2 - node->x1 + 1) * map.gridSize;
-      // Outer liner rectangle (border width 2)
-      SDL_Rect rect1 = {x, y, w, h};
-      SDL_Rect rect2 = {x + 1, y + 1, w - 2, h - 2};
-      renderDrawRect(rect1, Black);
-      renderDrawRect(rect2, Black);
-    };
-    mp->Nodes(c1);
-  }
-}
-void Visualizer::renderGates() {
-  auto mp = getCurrentQuadtreeMapByAgent();
-  if (mp != nullptr) {
-    // Gates.
-    qdpf::internal::GateVisitor visitor = [this, &mp](const qdpf::internal::Gate* gate) {
-      auto [x1, y1] = mp->UnpackXY(gate->a);
-      int x3 = y1 * map.gridSize + 1;
-      int y3 = x1 * map.gridSize + 1;
-      SDL_Rect rect = {x3, y3, map.gridSize - 2, map.gridSize - 2};
-      renderFillRect(rect, LightPurple);
-    };
-
-    mp->Gates(visitor);
-  }
-}
-
-void Visualizer::renderPathfindingDispatch() {
-  switch (pathfinderFlag) {
-    case PathFinderFlag::AStar:
-      renderPathfindingAStar();
-      return;
-    case PathFinderFlag::FlowField:
-      renderPathfindingFlowField();
-      return;
-  }
-}
-
-void Visualizer::renderHighlightedNodes() {
-  // render highlighted nodes with background at first.
-  switch (state) {
-    case State::AStarNodePathComputed:
-      [[fallthrough]];
-    case State::AStarGatePathComputed:
-      [[fallthrough]];
-    case State::AStarFinalPathComputed:
-      renderHighlightedNodesAstar();
-      break;
-    case State::FlowFieldNodeLevelComputed:
-      renderHighlightedNodesFlowField();
-      [[fallthrough]];
-    case State::FlowFieldGateLevelComputed:
-      [[fallthrough]];
-    case State::FlowFieldFinalLevelComputed:
-      renderHighlightedNodesFlowField();
-      break;
-    default:
-      break;  // avoiding warning
-  }
-}
-
-void Visualizer::renderHighlightedNodesAstar() {
-  for (auto node : astar.nodePath) {
-    int x1 = node->x1, y1 = node->y1, x2 = node->x2, y2 = node->y2;
-    int h = (x2 - x1 + 1) * map.gridSize;
-    int w = (y2 - y1 + 1) * map.gridSize;
-    int x = y1 * map.gridSize;
-    int y = x1 * map.gridSize;
-    SDL_Rect rect = {x, y, w, h};
-    SDL_Rect inner = {x + 1, y + 1, w - 2, h - 2};
-    renderFillRect(inner, LightOrange);
-  }
-}
-
-void Visualizer::renderHighlightedNodesFlowField() {
-  for (auto [node, nextNode, cost] : flowfield.nodeFlowField) {
-    // highlight node with background orange
-    int x1 = node->x1, y1 = node->y1, x2 = node->x2, y2 = node->y2;
-    int h = (x2 - x1 + 1) * map.gridSize;
-    int w = (y2 - y1 + 1) * map.gridSize;
-    int x = y1 * map.gridSize;
-    int y = x1 * map.gridSize;
-    SDL_Rect rect = {x, y, w, h};
-    SDL_Rect inner = {x + 1, y + 1, w - 2, h - 2};
-    renderFillRect(inner, LightOrange);
-  }
-}
-
-void Visualizer::renderPathfindingAStar() {
-  int agentSizeInPixels = agent.size * map.gridSize / COST_UNIT;
-
-  switch (state) {
-    case State::AStarWaitTarget:
-      renderFillCell(astar.x1, astar.y1, Green);  // start
-      break;
-    case State::AStarWaitCompution:
-      [[fallthrough]];
-    case State::AStarNodePathComputed:
-      renderFillCell(astar.x1, astar.y1, Green);  // start
-      renderFillCell(astar.x2, astar.y2, Green);  // target
-      // NOTE: nodepath already render in renderHighlightedNodesAstar.
-      break;
-    case State::AStarGatePathComputed:
-      // draw gate route cells.
-      // start and target are included in gate path.
-      for (const auto [x, y] : astar.gatePath) {
-        renderFillCell(x, y, Green);
-      }
-      break;
-    case State::AStarFinalPathComputed:  //
-      for (const auto [x, y] : astar.finalPath) {
-        renderFillAgent(x, y);
-      }
-      break;
-    default:
-      return;  // do nothing
-  }
-}
-
-void Visualizer::renderPathfindingFlowField() {
-  auto drawQrangeRectangle = [this]() {
-    // render the qrange rect
-    SDL_Rect qrange{
-        flowfield.qrange.y1 * map.gridSize,
-        flowfield.qrange.x1 * map.gridSize,
-        (flowfield.qrange.y2 - flowfield.qrange.y1 + 1) * map.gridSize,
-        (flowfield.qrange.x2 - flowfield.qrange.x1 + 1) * map.gridSize,
-    };
-    SDL_Rect inner{qrange.x + 1, qrange.y + 1, qrange.w - 2, qrange.h - 2};
-    renderDrawRect(qrange, Green);
-    renderDrawRect(inner, Green);
-  };
-
-  switch (state) {
-    case State::FlowFieldWaitQrangeRightBottom:
-      renderFillCell(flowfield.qrange.x1, flowfield.qrange.y1, Green);  // qrange.left-top
-      return;
-    case State::FlowFieldWaitTarget:
-      drawQrangeRectangle();
-      renderFillCell(flowfield.qrange.x1, flowfield.qrange.y1, Green);  // qrange.left-top
-      renderFillCell(flowfield.qrange.x2, flowfield.qrange.y2, Green);  // qrange.right-bottom
-      return;
-    case State::FlowFieldWaitCompution:
-      [[fallthrough]];
-    case State::FlowFieldNodeLevelComputed:
-      // NOTE the highlighting of the node field are already done in
-      // renderHighlightedNodesFlowField
-      drawQrangeRectangle();
-      renderFillCell(flowfield.qrange.x1, flowfield.qrange.y1, Green);  // qrange.left-top
-      renderFillCell(flowfield.qrange.x2, flowfield.qrange.y2, Green);  // qrange.right-bottom
-      renderFillCell(flowfield.x2, flowfield.y2, Green);                // target
-      return;
-    case State::FlowFieldGateLevelComputed:
-      drawQrangeRectangle();
-      renderPathFindingFlowFieldGateField();
-      return;
-    case State::FlowFieldFinalLevelComputed:
-      drawQrangeRectangle();
-      renderFillCell(flowfield.x2, flowfield.y2, Green);  // target
-      renderPathFindingFlowFieldFinalField();
-      return;
-    default:
-
-      return;  // do nothing
-  }
-}
-
-void Visualizer::renderPathFindingFlowFieldGateField() {
-  // draw gate cells on the flowfield.
-  for (int i = 0; i < flowfield.gateFlowField.size(); ++i) {
-    auto& [gate, nextNext, cost] = flowfield.gateFlowField[i];
-    auto [x, y] = gate;
-    SDL_Rect cell{y * map.gridSize + 1, x * map.gridSize + 1, map.gridSize - 2, map.gridSize - 2};
-    renderFillRect(cell, Green);
-  }
-}
-
-void Visualizer::renderPathFindingFlowFieldFinalField() {
-  if (flowfield.testPath.size()) {
-    for (auto [x, y] : flowfield.testPath) {
-      renderFillAgent(x, y);
-    }
-  }
-  for (int i = 0; i < flowfield.finalFlowField.size(); ++i) {
-    auto& [u, v, cost] = flowfield.finalFlowField[i];
-    auto [x, y] = u;
-    auto [x1, y1] = v;
-    if (u == v) continue;
-    if (x >= flowfield.qrange.x1 && x <= flowfield.qrange.x2 && y >= flowfield.qrange.y1 &&
-        y <= flowfield.qrange.y2) {
-      int dx = (x1 - x), dy = (y1 - y);
-      int d = (dx + 1) * 3 + (dy + 1);
-      if (!(d >= 0 && d <= 8 && d != 4)) continue;
-      int k = ARROWS_DIRECTIONS[d];
-      int w = arrows.w[k], h = arrows.h, offset = arrows.offset[k];
-      w = std::min(w, map.gridSize);
-      h = std::min(h, map.gridSize);
-      SDL_Rect dst = {y * map.gridSize + std::max(0, map.gridSize / 2 - w / 2),
-                      x * map.gridSize + std::max(0, map.gridSize / 2 - h / 2), w, h};
-      SDL_Rect src = {offset, 0, w, h};
-      renderCopy(arrows.texture, src, dst);
-    }
-  }
-}
-
-void Visualizer::renderDrawRect(const SDL_Rect& rect, const SDL_Color& color) {
-  SDL_Rect overlap;
-  if (cropRectByCamera(rect, overlap)) {
-    SDL_SetRenderDrawColor(renderer, color.r, color.g, color.b, color.a);
-    SDL_RenderDrawRect(renderer, &overlap);
-  }
-}
-
-void Visualizer::renderFillRect(const SDL_Rect& rect, const SDL_Color& color) {
-  SDL_Rect overlap;
-  if (cropRectByCamera(rect, overlap)) {
-    SDL_SetRenderDrawColor(renderer, color.r, color.g, color.b, color.a);
-    SDL_RenderFillRect(renderer, &overlap);
-  }
-}
-
-void Visualizer::renderDrawLine(int x1, int y1, int x2, int y2, const SDL_Color& color) {
-  // TODO: how to crop?
-  x1 -= camera->x;
-  x2 -= camera->x;
-  y1 -= camera->y;
-  y2 -= camera->y;
-  SDL_SetRenderDrawColor(renderer, color.r, color.g, color.b, color.a);
-  SDL_RenderDrawLine(renderer, x1, y1, x2, y2);
-}
-
-void Visualizer::renderCopy(SDL_Texture* texture, const SDL_Rect& src, const SDL_Rect& dst) {
-  SDL_Rect dstOverlap;
-  if (!cropRectByCamera(dst, dstOverlap, false)) return;
-  // find the crop region in src
-  SDL_Rect srcOverlap = {src.x + dstOverlap.x - dst.x, src.y + dstOverlap.y - dst.y, dstOverlap.w,
-                         dstOverlap.h};
-  dstOverlap.x -= camera->x;
-  dstOverlap.y -= camera->y;
-  SDL_RenderCopy(renderer, texture, &srcOverlap, &dstOverlap);
-}
-
-void Visualizer::renderFillCell(int x, int y, const SDL_Color& color) {
-  SDL_Rect cell{y * map.gridSize + 1, x * map.gridSize + 1, map.gridSize - 2, map.gridSize - 2};
-  renderFillRect(cell, color);
-}
-
-void Visualizer::renderFillAgent(int x, int y) {
-  int agentSizeInPixels = agent.size * map.gridSize / COST_UNIT;
-  SDL_Rect outer{y * map.gridSize, x * map.gridSize, agentSizeInPixels, agentSizeInPixels};
-  SDL_Rect inner{outer.x + 1, outer.y + 1, outer.w - 2, outer.h - 2};
-  renderDrawRect(outer, Black);
-  renderFillRect(inner, Green);
-}
-
-// Crop given rect by camera, and converts to the coordinates relative to the camera's left-top
-// corner. The results are stored into overlap. Returns false if no overlaping.
-bool Visualizer::cropRectByCamera(const SDL_Rect& rect, SDL_Rect& overlap,
-                                  bool marginToCameraCoordinates) {
-  qdpf::Rectangle c{camera->x, camera->y, camera->x + camera->w - 1, camera->y + camera->h - 1};
-  qdpf::Rectangle a{rect.x, rect.y, rect.x + rect.w - 1, rect.y + rect.h - 1};
-  qdpf::Rectangle d;
-  auto b = qdpf::internal::GetOverlap(c, a, d);
-  if (!b) return false;
-  overlap.x = d.x1, overlap.y = d.y1;
-  overlap.w = d.x2 - d.x1 + 1, overlap.h = d.y2 - d.y1 + 1;
-  if (marginToCameraCoordinates) {
-    overlap.x -= camera->x;
-    overlap.y -= camera->y;
-  }
-  return true;
-}
-
-// reset path finder compution results.
-void Visualizer::reset() {
-  state = State::Idle;
-  astar.Reset();
-  flowfield.Reset();
-  changeTo = Terrain::Building;
-  changingTerrainCells.clear();
-  showClearAllTerrainsConfirm = false;
-  setMessageHint("Reset done!", ImGreen);
-}
-
-void Visualizer::setMessageHint(std::string_view message, const ImVec4& color) {
-  messageHint = message;
-  messageHintColor = color;
-}
-
-void Visualizer::handleStartDrawBuildings() {
-  reset();
-  state = State::DrawingBuildings;
-  changeTo = Terrain::Building;
-  setMessageHint("click or drag mouse to draw buildings!", ImGreen);
-}
-
-void Visualizer::handleStartDrawWater() {
-  reset();
-  state = State::DrawingWaters;
-  changeTo = Terrain::Water;
-  setMessageHint("click or drag mouse to draw water!", ImGreen);
-}
-
-void Visualizer::pushTerrainChanges(const Cell& cell) {
-  auto [x, y] = cell;
-  // invert between land and changeTo.
-  auto to = (map.grids[x][y] == Terrain::Land) ? changeTo : Terrain::Land;
-  if (x >= 0 && x < map.h && y >= 0 && y < map.w && !map.changes[x][y]) {
-    changingTerrainCells.push_back(cell);
-    map.WantChangeTerrain(cell, to);
-  }
-}
-
-void Visualizer::applyTerrainChanges() {
-  if (changingTerrainCells.empty()) return;
-  map.ApplyChangeTerrain(changingTerrainCells);
-  changingTerrainCells.clear();
-}
-
-void Visualizer::handleSwitchPathFinderHandler(PathFinderFlag to) {
-  if (to == pathfinderFlag) return;
-  reset();
-  pathfinderFlag = to;
-  setMessageHint("Old states reset done and pathfinder switched.", ImGreen);
-}
-
-void Visualizer::handleChangeAgentSize(int to) {
-  if (agent.size == to) return;
-  reset();
-  auto oldAgentSize = agent.size;
-  agent.size = to;
-  if (getCurrentQuadtreeMapByAgent() == nullptr) {
-    setMessageHint("Failed to change agent size, stay unchanged!", ImRed);
-    agent.size = oldAgentSize;
-  } else
-    setMessageHint("Agent size changed (current quadtree map switched)!", ImGreen);
-}
-
-void Visualizer::handleChangeAgentCompability(int to) {
-  if (agent.capability == to) return;
-  reset();
-  auto oldCapability = agent.capability;
-  agent.capability = to;
-  if (getCurrentQuadtreeMapByAgent() == nullptr) {
-    setMessageHint("Failed to change agent capability, stay unchanged!", ImRed);
-    agent.size = oldCapability;
-  } else
-    setMessageHint("Agent capability changed (current quadtree map switched)!", ImGreen);
-}
-
-void Visualizer::handleClearAllTerrains() {
-  reset();
-  map.ClearAllTerrains();
-  setMessageHint("Map is cleared.", ImGreen);
-}
-
-// Returns the pointer the internal quadtree map of which current agent is using.
-const qdpf::internal::QuadtreeMap* Visualizer::getCurrentQuadtreeMapByAgent() const {
-  return map.qmx->Get(agent.size, agent.capability);
-}
-
-// returns cell at the position for (x,y) in the camera.
-Cell Visualizer::getCellAtPixelPosition(int x, int y) const {
-  return {(y + camera->y) / map.gridSize, (x + camera->x) / map.gridSize};
-}
-
-void Visualizer::handleAstarInputBegin() {
-  if (pathfinderFlag != PathFinderFlag::AStar) {
-    pathfinderFlag = PathFinderFlag::AStar;
-  }
-  if (state != State::Idle) reset();
-  state = State::AStarWaitStart;
-  setMessageHint("A*: waiting to click a start cell", ImGreen);
-}
-
-void Visualizer::computeAstarNodePath() {
-  if (state != State::AStarWaitCompution) {
-    setMessageHint("invalid state", ImRed);
-    return;
-  }
-  if (0 != astar.ResetPf(agent.size, agent.capability)) {
-    setMessageHint("internal error: astar reset failure", ImRed);
-    return;
-  }
-  std::chrono::high_resolution_clock::time_point startAt, endAt;
-
-  startAt = std::chrono::high_resolution_clock::now();
-  int cost = astar.pf->ComputeNodeRoutes();
-  endAt = std::chrono::high_resolution_clock::now();
-
-  state = State::AStarNodePathComputed;
-  if (cost == -1) {
-    setMessageHint("A*: unreachable!", ImRed);
-    return;
-  }
-
-  if (astar.nodePath.size()) astar.nodePath.clear();
-  qdpf::NodeVisitor visitor = [this](const qdpf::QdNode* node) { astar.nodePath.push_back(node); };
-  if (astar.pf->NodePathSize()) astar.pf->VisitComputedNodeRoutes(visitor);
-
-  setMessageHint(
-      fmt::format(
-          "A*: Node path computed! cost {}us ; Next we can click button < Compute Gate Path >.",
-          std::chrono::duration_cast<std::chrono::microseconds>(endAt - startAt).count()),
-      ImGreen);
-}
-
-void Visualizer::computeAstarGatePath() {
-  if (state != State::AStarWaitCompution && state != State::AStarNodePathComputed) {
-    setMessageHint("invalid state", ImRed);
-    return;
-  }
-  if (0 != astar.ResetPf(agent.size, agent.capability)) {
-    setMessageHint("internal error: astar reset failure", ImRed);
-    return;
-  }
-
-  std::chrono::high_resolution_clock::time_point startAt, endAt;
-
-  if (astar.gatePath.size()) astar.gatePath.clear();
-  bool useNodePath = astar.pf->NodePathSize() > 0;
-  qdpf::CellCollector collector = [this](int x, int y) { astar.gatePath.push_back({x, y}); };
-
-  startAt = std::chrono::high_resolution_clock::now();
-  int cost = astar.pf->ComputeGateRoutes(collector, useNodePath);
-  endAt = std::chrono::high_resolution_clock::now();
-
-  state = State::AStarGatePathComputed;
-  if (cost == -1) {
-    setMessageHint("A*: unreachable!", ImRed);
-    return;
-  }
-
-  setMessageHint(
-      fmt::format("A*: Gate path computed! useNodePath: {}  cost {}us ; Next we can click button "
-                  "< Compute Final Path >.",
-                  useNodePath,
-                  std::chrono::duration_cast<std::chrono::microseconds>(endAt - startAt).count()),
-      ImGreen);
-}
-
-void Visualizer::computeAstarFinalPath() {
-  if (state != State::AStarGatePathComputed) {
-    setMessageHint("invalid state", ImRed);
-    return;
-  }
-  if (astar.gatePath.empty()) {
-    setMessageHint("A*: empty gate route cells!", ImRed);
-    return;
-  }
-  if (astar.finalPath.size()) astar.finalPath.clear();
-
-  std::chrono::high_resolution_clock::time_point startAt, endAt;
-
-  qdpf::CellCollector collector = [this](int x, int y) {
-    if (astar.finalPath.size()) {
-      auto [x2, y2] = astar.finalPath.back();
-      if ((x2 == x && y2 == y)) return;
-    }
-    astar.finalPath.push_back({x, y});
-  };
-
-  startAt = std::chrono::high_resolution_clock::now();
-
-  auto [x, y] = astar.gatePath[0];
-  for (int i = 1; i < astar.gatePath.size(); i++) {
-    auto [x2, y2] = astar.gatePath[i];
-    qdpf::ComputeStraightLine(x, y, x2, y2, collector);
-    x = x2, y = y2;
-  }
-  state = State::AStarFinalPathComputed;
-  endAt = std::chrono::high_resolution_clock::now();
-
-  setMessageHint(
-      fmt::format(
-          "A*: final path computed! cost {}us ; Click button < Reset > to clear these results.",
-          std::chrono::duration_cast<std::chrono::microseconds>(endAt - startAt).count()),
-      ImGreen);
-}
-
-std::string StateToString(State state) {
-  switch (state) {
-    case State::Idle:
-      return "Idle";
-    case State::DrawingBuildings:
-      return "Changing Terrain (Building <=> Land)";
-    case State::DrawingWaters:
-      return "Changing Terrain (Water <=> Land)";
-    case State::AStarWaitStart:
-      return "(A*) Waiting Input Start Cell";
-    case State::AStarWaitTarget:
-      return "(A*) Waiting Input Target Cell";
-    case State::AStarWaitCompution:
-      return "(A*) Waiting Compution";
-    case State::AStarNodePathComputed:
-      return "(A*) Node Path Computed";
-    case State::AStarGatePathComputed:
-      return "(A*) Gate Path Computed";
-    case State::AStarFinalPathComputed:
-      return "(A*) Final Path Computed";
-    case State::FlowFieldWaitQrangeLeftTop:
-      return "(FlowField) Wait Input Query Range (left-top)";
-    case State::FlowFieldWaitQrangeRightBottom:
-      return "(FlowField) Wait Input Query Range (right-bottom)";
-    case State::FlowFieldWaitTarget:
-      return "(FlowField) Wait Input Target";
-    case State::FlowFieldWaitCompution:
-      return "(FlowField) Wait Compution";
-    case State::FlowFieldNodeLevelComputed:
-      return "(FlowField) Node Flow Field Computed";
-    case State::FlowFieldGateLevelComputed:
-      return "(FlowField) Gate Flow Field Computed";
-    case State::FlowFieldFinalLevelComputed:
-      return "(FlowField) Final Flow Field Computed";
-    default:
-      return "Unknown";
-  }
-}
-
-void Visualizer::handleFlowFieldInputQueryRangeBegin() {
-  if (pathfinderFlag != PathFinderFlag::FlowField) {
-    pathfinderFlag = PathFinderFlag::FlowField;
-  }
-  if (state != State::Idle) reset();
-  state = State::FlowFieldWaitQrangeLeftTop;
-  setMessageHint("FlowField: waiting to click a left-top cell", ImGreen);
-}
-
-void Visualizer::computeNodeFlowField() {
-  if (state != State::FlowFieldWaitCompution) {
-    setMessageHint("invalid state", ImRed);
-    return;
-  }
-  if (0 != flowfield.ResetPf(agent.size, agent.capability)) {
-    setMessageHint("internal error: flowfield reset failure", ImRed);
-    return;
-  }
-
-  std::chrono::high_resolution_clock::time_point startAt, endAt;
-
-  startAt = std::chrono::high_resolution_clock::now();
-  int ret = flowfield.pf->ComputeNodeFlowField();
-  endAt = std::chrono::high_resolution_clock::now();
-
-  state = State::FlowFieldNodeLevelComputed;
-  if (ret == -1) {
-    setMessageHint("FlowField: unreachable!", ImRed);
-    return;
-  }
-
-  if (flowfield.nodeFlowField.size()) flowfield.nodeFlowField.clear();
-
-  qdpf::NodeFlowFieldVisitor visitor = [this](const qdpf::QdNode* node,
-                                              const qdpf::QdNode* nextNode, int cost) {
-    if (node != nullptr) flowfield.nodeFlowField.push_back({node, nextNode, cost});
-  };
-  flowfield.pf->VisitComputedNodeFlowField(visitor);
-
-  // sort to draw from left-top to right-bottom.
-  auto cmp = [](const FlowFieldItem<const qdpf::QdNode*>& a,
-                const FlowFieldItem<const qdpf::QdNode*>& b) {
-    return ((a.current->x1) < (b.current->x1)) || ((a.current->y1) < (b.current->y1));
-  };
-  std::stable_sort(flowfield.nodeFlowField.begin(), flowfield.nodeFlowField.end(), cmp);
-
-  setMessageHint(
-      fmt::format("FlowField: Node flowfield computed! cost {}us ; Next we can click button < "
-                  "Compute Gate Flow Field  >.",
-                  std::chrono::duration_cast<std::chrono::microseconds>(endAt - startAt).count()),
-      ImGreen);
-}
-
-void Visualizer::computeGateFlowField() {
-  if (state != State::FlowFieldWaitCompution && state != State::FlowFieldNodeLevelComputed) {
-    setMessageHint("invalid state", ImRed);
-    return;
-  }
-
-  if (0 != flowfield.ResetPf(agent.size, agent.capability)) {
-    setMessageHint("internal error: flowfield reset failure", ImRed);
-    return;
-  }
-
-  std::chrono::high_resolution_clock::time_point startAt, endAt;
-
-  if (flowfield.gateFlowField.size()) flowfield.gateFlowField.clear();
-  bool useNodeFlowField = flowfield.nodeFlowField.size() > 0;
-
-  startAt = std::chrono::high_resolution_clock::now();
-  int ret = flowfield.pf->ComputeGateFlowField(useNodeFlowField);
-  endAt = std::chrono::high_resolution_clock::now();
-
-  state = State::FlowFieldGateLevelComputed;
-
-  if (-1 == ret) {
-    setMessageHint("FlowField: unreachable!", ImRed);
-    return;
-  }
-
-  qdpf::CellFlowFieldVisitor visitor = [this](int x, int y, int xNext, int yNext, int cost) {
-    flowfield.gateFlowField.push_back({{x, y}, {xNext, yNext}, cost});
-  };
-  flowfield.pf->VisitComputedGateFlowField(visitor);
-
-  // sort to draw from left-top to right-bottom
-  std::stable_sort(flowfield.gateFlowField.begin(), flowfield.gateFlowField.end(),
-                   [](const FlowFieldItem<Cell>& a, const FlowFieldItem<Cell>& b) {
-                     return a.current < b.current;
-                   });
-
-  setMessageHint(
-      fmt::format("Flowfield:: Gate flow field computed! useNodeFlowField: {}  cost {}us ; Next "
-                  "we can click button "
-                  "< Compute Final Flow Field >.",
-                  useNodeFlowField,
-                  std::chrono::duration_cast<std::chrono::microseconds>(endAt - startAt).count()),
-      ImGreen);
-}
-
-void Visualizer::computeFinalFlowField() {
-  if (state != State::FlowFieldGateLevelComputed) {
-    setMessageHint("invalid state", ImRed);
-    return;
-  }
-
-  std::chrono::high_resolution_clock::time_point startAt, endAt;
-
-  if (flowfield.finalFlowField.size()) flowfield.finalFlowField.clear();
-  if (flowfield.finalFlowNextMap.size()) flowfield.finalFlowNextMap.clear();
-
-  startAt = std::chrono::high_resolution_clock::now();
-  int ret = flowfield.pf->ComputeFinalFlowFieldInQueryRange();
-  endAt = std::chrono::high_resolution_clock::now();
-  state = State::FlowFieldFinalLevelComputed;
-
-  if (-1 == ret) {
-    setMessageHint("FlowField: unreachable!", ImRed);
-    return;
-  }
-
-  qdpf::CellFlowFieldVisitor visitor = [this](int x, int y, int xNext, int yNext, int cost) {
-    flowfield.finalFlowField.push_back({{x, y}, {xNext, yNext}, cost});
-    flowfield.finalFlowNextMap[{x, y}] = {xNext, yNext};
-  };
-
-  flowfield.pf->VisitComputedCellFlowFieldInQueryRange(visitor);
-
-  std::stable_sort(flowfield.finalFlowField.begin(), flowfield.finalFlowField.end(),
-                   [](const FlowFieldItem<Cell>& a, const FlowFieldItem<Cell>& b) {
-                     return a.current < b.current;
-                   });
-
-  setMessageHint(
-      fmt::format("Flowfield:: Final flow field computed!  cost {}us  ",
-                  std::chrono::duration_cast<std::chrono::microseconds>(endAt - startAt).count()),
-      ImGreen);
-}
-
-void Visualizer::handleLogics() {
-  switch (state) {
-    case State::FlowFieldFinalLevelComputed:
-      handlePlayFlowFieldTestPath();
-      break;
-    default:
-      break;
-  }
-}
-
-void Visualizer::handlePlayFlowFieldTestPath() {
-  // (x1,y1) is the start
-  int x1 = flowfield.x1, y1 = flowfield.y1;
-  if (x1 == -1 || y1 == -1) return;
-
-  // (x2,y2) is thet target
-  int x2 = flowfield.x2, y2 = flowfield.y2;
-
-  auto& p = flowfield.testPath;
-
-  if (p.empty()) {
-    p.push_back({x1, y1});
-    return;
-  }
-
-  // get current position
-  auto [x3, y3] = p.back();
-
-  if (x3 == x2 && y3 == y2) {
-    // arrived the target.
-    // back to start;
-    p.clear();
-    p.push_back({x1, y1});
-    return;
-  }
-
-  // Is inside the rect?
-  if ((x3 >= flowfield.qrange.x1 && x3 <= flowfield.qrange.x2 && y3 >= flowfield.qrange.y1 &&
-       y3 <= flowfield.qrange.y2)) {
-    // get next from the final flow field.
-    p.push_back(flowfield.finalFlowNextMap[{x3, y3}]);
-  }
 }
 
 int ParseCommandlineOptions(int argc, char* argv[]) {
